Fixes some compiler warnings

+++ src/hugin1/hugin/PreviewControlPointTool.cpp
@@ -59,7 +59,7 @@ void PreviewControlPointTool::AfterDrawImagesEvent()
     // now draw each control point in turn:
     helper->GetViewStatePtr()->GetTextureManager()->DisableTexture();
     glColor3f(1.0, 0.5, 0.0);
-    int cp_count = control_points.size();
+    size_t cp_count = control_points.size();
     for (size_t cp_index = 0; cp_index < cp_count; cp_index++)
     {
         const HuginBase::ControlPoint &cp = control_points[cp_index];
